backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hirokts", email = "9495043+hirokts@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ag-ui-langgraph>=0.0.21",
    "copilotkit>=0.1.72",
    "fastapi>=0.115.14",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "uvicorn>=0.38.0",
]

[project.scripts]
backend = "backend:main"
serve = "backend:dev"

[build-system]
requires = ["uv_build>=0.9.12,<0.10.0"]
build-backend = "uv_build"
